handled size comparison warning

+++ src/okvm.c
@@ -87,13 +87,13 @@ static void handle_opcode(OkVM* vm, unsigned char opcode, unsigned char argsize)
       break;
     case 6: // str
       addr = (size_t) stack_popn(&(vm->dst), VM_WORD_SIZE);
-      for (size_t i = 0; i < argsize + 1; i++) {
+      for (int i = 0; i < argsize + 1; i++) {
         vm->ram[addr + i] = stack_pop(&(vm->dst)); 
       }
       break;
     case 7: // lod
       addr = (size_t) stack_popn(&(vm->dst), VM_WORD_SIZE);
-      for (size_t i = 0; i < argsize + 1; i++) {
+      for (int i = 0; i < argsize + 1; i++) {
         stack_push(&(vm->dst), vm->ram[addr + i]);
       }
       break;
@@ -117,7 +117,7 @@ static void handle_opcode(OkVM* vm, unsigned char opcode, unsigned char argsize)
       vm->pc = (size_t) stack_popn(&(vm->dst), argsize + 1);
       break;
     case 13: // lit
-      for (size_t i = 0; i < argsize + 1; i++) {
+      for (int i = 0; i < argsize + 1; i++) {
         stack_push(&(vm->dst), fetch(vm));
       }
       break;
